+++ apps/ethereum/erc721/components/WriteFunctions/CreateInstance.tsx
@@ -1,0 +1,12 @@
+
+import { useState } from 'react';
+import * as Accordion from '@radix-ui/react-accordion';
+import { useEthereum } from '@decentology/hyperverse-ethereum';
+// import { useToken } from '../../source/useToken';
+import { Box, Item, TriggerContainer, Trigger, Parameters, Input, Content, Button } from './WriteComponents'
+
+const CreateInstance = () => {
+  return <div></div>
+}
+
+export default CreateInstance;