+++ packages/hyperverse-ethereum-erc721/source/useERC721.ts
@@ -1,36 +1,36 @@
 import { useState, useEffect, useCallback } from 'react';
 import { useQuery, useMutation, useQueryClient, UseMutationOptions } from 'react-query';
 import { ethers } from 'ethers';
+import ERC721FactoryABI from '../artifacts/contracts/ExampleNFTFactory.sol/ExampleNFTFactory.json';
+import ERC721ABI from '../artifacts/contracts/ExampleNFT.sol/ExampleNFT.json';
 import { useEthereum } from '@decentology/hyperverse-ethereum';
-import { ContractABI, CONTRACT_TESTNET_ADDRESS } from './Provider';
-import { useEvent } from 'react-use';
 import { useStorage } from '@decentology/hyperverse-storage-skynet';
 import { createContainer, useContainer } from 'unstated-next';
 
 type ContractState = ethers.Contract;
 
-type MetaData = {
-	name: string;
-	description: string;
-	image: string;
-};
+/*
+ExampleNFT deployed to: 0x0af78495a25D77D25F51D4F797F473521f29e07B
+ExampleNFTFactory deployed to: 0xe5d761311212ABF55c9C6eb6d80eAF804F213d72
+*/
+
+const EXAMPLENFT_FACTORY_ADDRESS = '0xe5d761311212ABF55c9C6eb6d80eAF804F213d72';
+const TENANT_ADDRESS = '0xD847C7408c48b6b6720CCa75eB30a93acbF5163D';
 
 function ERC721State(initialState: { tenantId: string } = { tenantId: '' }) {
 	const { tenantId } = initialState;
-	const queryClient = useQueryClient();
 	const { address, web3Provider, provider, connect } = useEthereum();
-	const { clientUrl } = useStorage();
-	const [contract, setExampleNFTContract] = useState<ContractState>(
-		new ethers.Contract(CONTRACT_TESTNET_ADDRESS, ContractABI, provider) as ContractState
+	const [contract, setContract] = useState<ContractState>(
+		new ethers.Contract(EXAMPLENFT_FACTORY_ADDRESS, ERC721FactoryABI.abi, provider) as ContractState
 	);
-	const { uploadFile } = useStorage();
+
 	const setup = useCallback(async () => {
 		const signer = await web3Provider?.getSigner();
 		if (signer && contract) {
 			const ctr = contract.connect(signer) as ContractState;
-			setExampleNFTContract(ctr);
+			setContract(ctr);
 		}
-	}, [web3Provider]);
+	}, [contract, web3Provider]);
 
 	const errors = (err: any) => {
 		if (!contract?.signer) {
@@ -41,21 +41,14 @@ function ERC721State(initialState: { tenantId: string } = { tenantId: '' }) {
 			throw new Error('You rejected the transaction!');
 		}
 
-		if (err.message.includes('User is already in a Tribe!')) {
-			throw new Error('You are already in a tribe!');
-		}
-
 		throw err;
-		// throw new Error("Something went wrong!");
 	};
 
 	useEffect(() => {
 		if (web3Provider) {
 			setup();
 		}
-	}, [web3Provider]);
-
-	// How do I connect all these calls to the proxy?
+	}, [setup, web3Provider]);
 
 	const createInstance = useCallback(async (name: string, symbol: string) => {
 		try {
@@ -65,6 +58,17 @@ function ERC721State(initialState: { tenantId: string } = { tenantId: '' }) {
 			errors(err);
 			throw err;
 		}
+	}, [contract]);
+
+	const getProxy = useCallback(async (account: string) => {
+		try {
+			const proxyAccount = await contract.getProxy(account);
+			return proxyAccount;
+		} catch (err) {
+			errors(err);
+			throw err;
+		}
+		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [contract]);
 
 	const mintNFT = useCallback(async () => {
@@ -77,27 +81,29 @@ function ERC721State(initialState: { tenantId: string } = { tenantId: '' }) {
 		}
 	}, [contract]);
 
-	const getBalanceOf = useCallback(async (address) => {
+	const getBalance = useCallback(async (address) => {
 		try {
-			const balance = await contract.getProxy(tenantId).balanceOf(address);
+			const proxyAddress = await contract.getProxy(TENANT_ADDRESS)
+			const proxyContract = new ethers.Contract(proxyAddress, ERC721ABI.abi, provider)
 
+			const balance = await proxyContract.balanceOf(address);
 			return balance.toNumber();
 		} catch (err) {
 			throw err;
 		}
-	}, [contract]);
+	}, [contract, provider]);
 
 	return {
 		tenantId,
 		contract,
 		NewInstance: (
 			options?: Omit<UseMutationOptions<unknown, unknown, unknown, unknown>, 'mutationFn'>
-		) => useMutation((name: string, symbol: string) => createInstance(name, symbol), options),
+		) => useMutation(({ name, symbol }) => createInstance(name, symbol), options),
 		MintNFT: (
 			options?: Omit<UseMutationOptions<unknown, unknown, void, unknown>, 'mutationFn'>
 		) => useMutation(() => mintNFT(), options),
-		BalanceOf: () =>
-			useQuery(['getTribeId', contract?.address], (address) => getBalanceOf(address), {
+		GetBalance: () =>
+			useQuery(['getBalance', address, contract?.address], (address) => getBalance(address), {
 				enabled: !!contract?.address,
 				retry: false,
 			}),