+++ packages/hyperverse-algorand/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @decentology/hyperverse-algorand
 
+## 1.0.8
+
+### Patch Changes
+
+-   9df884a: Updated package deps versions for semver
+-   Updated dependencies [9df884a]
+-   Updated dependencies
+    -   @decentology/hyperverse@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes