+++ packages/hyperverse-metis-tribes/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @decentology/hyperverse-ethereum-tribes
 
+## 1.0.9
+
+### Patch Changes
+
+-   Fixed version pinning on metis
+-   Updated dependencies
+    -   @decentology/hyperverse-metis@1.0.9
+
+## 1.0.8
+
+### Patch Changes
+
+-   Milestone update. Added support for multi chain swapping
+-   Updated dependencies [9df884a]
+-   Updated dependencies
+    -   @decentology/hyperverse@1.0.8
+    -   @decentology/hyperverse-storage-skynet@1.0.7
+    -   @decentology/hyperverse-metis@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes